+++ ff-core/lib/FF/Options.hs
@@ -17,7 +17,7 @@ module FF.Options
     Options (..),
     Search (..),
     Shuffle (..),
-    Tags(..),
+    Tags (..),
     Track (..),
     assignToMaybe,
     parseOptions,
@@ -37,6 +37,7 @@ import qualified FF.Types
 import Options.Applicative
   ( (<**>),
     Completer,
+    InfoMod,
     ParseError (ShowHelpText),
     Parser,
     ParserInfo,
@@ -49,6 +50,7 @@ import Options.Applicative
     defaultPrefs,
     flag,
     flag',
+    footer,
     fullDesc,
     help,
     helper,
@@ -129,7 +131,8 @@ assignToMaybe = \case
 data Agenda
   = Agenda
       { limit :: Maybe Limit,
-        tags :: Tags
+        tags :: Tags,
+        withoutTags :: Set Text
       }
 
 data Tags = Tags (Set Text) | NoTags
@@ -164,7 +167,8 @@ data Search
         inContacts :: Bool,
         status :: Status,
         limit :: Maybe Limit,
-        tags :: Tags
+        tags :: Tags,
+        withoutTags :: Set Text
       }
 
 parseOptions :: Maybe StorageFS.Handle -> IO Options
@@ -214,7 +218,13 @@ parser h =
     iCmdContact = i cmdContact "show contacts"
     iCmdDelete = i cmdDelete "delete a task"
     iCmdDone = i cmdDone "mark a task done (archive)"
-    iCmdEdit = i cmdEdit "edit a task or a note"
+    iCmdEdit =
+      i_
+        cmdEdit
+        "edit a task or a note, using command from environment variable EDITOR\
+        \ or program `editor`"
+        $ footer
+          "Examples for EDITOR: 'code --wait', 'emacs', 'micro', 'nano', 'vim'"
     iCmdNew = i cmdNew "synonym for `add`"
     iCmdPostpone = i cmdPostpone "make a task start later"
     iCmdSearch = i cmdSearch "search for notes with the given text"
@@ -244,17 +254,17 @@ parser h =
     wiki = switch $ long "wiki" <> short 'w' <> help "Handle wiki note"
     briefOption =
       switch $ long "brief" <> short 'b' <> help "List only note titles and ids"
-    agenda = Agenda <$> optional limitOption <*> filterTags
+    agenda = Agenda <$> optional limitOption <*> filterTags <*> withoutTagsOption
     filterTags = filterByNoTags <|> Tags <$> filterByTags
     track = Track <$> dryRunOption <*> optional repo <*> optional limitOption
     dryRunOption =
-      switch
-        $ long "dry-run"
+      switch $
+        long "dry-run"
           <> short 'd'
           <> help "List only issues, don't set up tracking"
     repo =
-      strOption
-        $ long "repo"
+      strOption $
+        long "repo"
           <> short 'r'
           <> metavar "USER/REPO"
           <> help "User or organization/repository"
@@ -266,10 +276,12 @@ parser h =
           Add <$> strArgument (metavar "CONTACT_NAME" <> help "contact name")
         pDelete =
           Delete
-            <$> argument readDocId
-                  ( metavar "CONTACT_ID" <> help "contact id"
-                      <> completer completeContactIds
-                  )
+            <$> argument
+              readDocId
+              ( metavar "CONTACT_ID"
+                  <> help "contact id"
+                  <> completer completeContactIds
+              )
     new =
       New
         <$> noteTextArgument
@@ -296,30 +308,35 @@ parser h =
         <*> searchA
         <*> optional limitOption
         <*> filterTags
+        <*> withoutTagsOption
     searchT = switch $ long "tasks" <> short 't' <> help "Search among tasks"
     searchW = switch $ long "wiki" <> short 'w' <> help "Search among wiki"
     searchC =
       switch $ long "contacts" <> short 'c' <> help "Search among contacts"
     searchA =
-      flag Active Archived
-        $ long "archived" <> short 'a' <> help "Search among archived"
+      flag Active Archived $
+        long "archived" <> short 'a' <> help "Search among archived"
     noteid =
-      argument readDocId
-        $ metavar "ID" <> help "note id" <> completer completeNoteIds
+      argument readDocId $
+        metavar "ID" <> help "note id" <> completer completeNoteIds
     noteTextArgument = strArgument $ metavar "TEXT" <> help "Note's text"
     filterByTags =
-      fmap Set.fromList $ many $ strOption
-        $ long "tag" <> metavar "TAG" <> help "Filter by tag"
-    filterByNoTags = flag' NoTags $
-        long "no-tags"
+      fmap Set.fromList $ many $ strOption $
+        long "tag" <> metavar "TAG" <> help "Filter by tag"
+    filterByNoTags =
+      flag' NoTags $
+        long "no-tag"
           <> short 'n'
-          <> help "Filter items without tags"
+          <> help "Filter items that has no tags"
     addTagsOption =
-      fmap Set.fromList $ many $ strOption
-        $ long "tag" <> metavar "TAG" <> help "Add tags"
+      fmap Set.fromList $ many $ strOption $
+        long "tag" <> metavar "TAG" <> help "Add tag"
     deleteTagsOption =
-      fmap Set.fromList $ many $ strOption
-        $ long "delete-tag" <> short 'd' <> metavar "TAG" <> help "Delete a tag"
+      fmap Set.fromList $ many $ strOption $
+        long "delete-tag" <> short 'd' <> metavar "TAG" <> help "Delete tag"
+    withoutTagsOption =
+      fmap Set.fromList $ many $ strOption $
+        long "without-tag" <> metavar "TAG" <> help "Filter items without tag"
     endDateOption = dateOption $ long "end" <> short 'e' <> help "end date"
     deltaOption = option auto $
       long "repeat" <> short 'r' <> help "Number of days"
@@ -334,14 +351,14 @@ parser h =
         <|> flag' Clear (long "end-clear" <> help "clear end date")
     dateOption m = option auto $ metavar "DATE" <> m
     customDirOption =
-      optional $ strOption
-        $ long "data-dir"
+      optional $ strOption $
+        long "data-dir"
           <> short 'C'
           <> metavar "DIRECTORY"
           <> help "Path to the data dir"
     cmdConfig =
-      fmap CmdConfig . optional . subparser
-        $ command "dataDir" iDataDir <> command "ui" iUi
+      fmap CmdConfig . optional . subparser $
+        command "dataDir" iDataDir <> command "ui" iUi
       where
         iDataDir = i pDataDir "the database directory"
         pDataDir = ConfigDataDir <$> optional (pJust <|> pYandexDisk)
@@ -354,10 +371,10 @@ parser h =
                 (long "yandex-disk" <> short 'y' <> help "detect Yandex.Disk")
         iUi = i ui "UI tweaks"
         ui =
-          fmap ConfigUI . optional
-            $ flag'
-                Shuffle
-                (long "shuffle" <> help "shuffle notes in section")
+          fmap ConfigUI . optional $
+            flag'
+              Shuffle
+              (long "shuffle" <> help "shuffle notes in section")
               <|> flag' Sort (long "sort" <> help "sort notes in section")
     version =
       flag'
@@ -366,22 +383,25 @@ parser h =
     completeNoteIds = docIdCompleter @Note
     completeContactIds = docIdCompleter @FF.Types.Contact
     docIdCompleter :: forall a. Collection a => Completer
-    docIdCompleter = listIOCompleter
-      $ map unDocId <$> case h of
+    docIdCompleter = listIOCompleter $
+      map unDocId <$> case h of
         Nothing -> pure []
         Just h' -> runStorage h' (getDocuments @_ @a)
     unDocId (DocId name) = name
     readDocId = DocId <$> str
 
 i :: Parser a -> String -> ParserInfo a
-i prsr desc = info (prsr <**> helper) $ fullDesc <> progDesc desc
+i prsr desc = i_ prsr desc mempty
+
+i_ :: Parser a -> String -> InfoMod a -> ParserInfo a
+i_ prsr desc m = info (prsr <**> helper) $ fullDesc <> progDesc desc <> m
 
 parserInfo :: Maybe StorageFS.Handle -> ParserInfo Options
 parserInfo h = i (parser h) "A note taker and task tracker"
 
 showHelp :: String
 showHelp =
-  fst
-    $ renderFailure
-        (parserFailure prefs (parserInfo Nothing) ShowHelpText mempty)
-        ""
+  fst $
+    renderFailure
+      (parserFailure prefs (parserInfo Nothing) ShowHelpText mempty)
+      ""