+++ ff-core/lib/FF/UI.hs
@@ -10,12 +10,14 @@ module FF.UI (
     prettyContactSample,
     prettyNote,
     prettyNoteList,
+    prettyPath,
     prettyTagsList,
     prettyTaskSections,
     prettyTasksWikisContacts,
     prettyWikiSample,
     sampleLabel,
-    withHeader
+    withHeader,
+    (<//>)
 ) where
 
 import           Data.Char (isSpace)
@@ -27,7 +29,7 @@ import           Data.Semigroup ((<>))
 import           Data.Set (Set)
 import           Data.Text (Text)
 import qualified Data.Text as Text
-import           Data.Text.Prettyprint.Doc (Doc, annotate, fillSep, hang,
+import           Data.Text.Prettyprint.Doc (Doc, annotate, fillSep, hang, line,
                                             indent, pretty, sep, space, viaShow,
                                             vsep, (<+>))
 import           Data.Text.Prettyprint.Doc.Render.Terminal (AnsiStyle,
@@ -35,11 +37,11 @@ import           Data.Text.Prettyprint.Doc.Render.Terminal (AnsiStyle,
                                                             color)
 import           Data.Time (Day)
 import           FF (fromRgaM)
+import           FF.Options (Tags (..))
 import           FF.Types (Contact (..), ContactSample, Entity (..), EntityDoc,
                            EntityView, ModeMap, Note (..), NoteSample,
                            NoteStatus (Wiki), Sample (..), TaskMode (..),
                            Track (..), View (..), omitted)
-import           FF.Options (Tags(..))
 import           RON.Storage.Backend (DocId (DocId))
 
 -- | Header with fixed yellow color.
@@ -50,8 +52,15 @@ withHeader header value =
 indentation :: Int
 indentation = 2
 
+(<//>) :: Doc ann -> Doc ann -> Doc ann
+a <//> b = a <> line <> line <> b
+infixr 6 <//>
+
 prettyDocId :: DocId a -> Doc ann
 prettyDocId (DocId name) = pretty name
+
+prettyPath :: Maybe FilePath -> Doc AnsiStyle
+prettyPath path = yellow "Database:" <+> pretty path
 
 prettyTasksWikisContacts
     :: Bool                      -- ^ is output brief